requirements.txt
amqp==2.2.1
astroid==1.5.3
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
dj-database-url==0.5.0
Django==2.0.4
# django-allauth==0.38.0
django-celery-beat==1.1.1
django-environ==0.4.5
# django-oauth-toolkit==1.2.0
# django-rest-auth==0.9.3
# django-slack-oauth==1.4.3
django-webpack-loader==0.6.0
djangorestframework==3.8.2
enum34==1.1.6
facepy==1.0.9
gunicorn==19.7.1
honcho==1.0.1
idna==2.6
isort==4.2.15
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.3
parsedatetime==2.4
Pillow==4.2.1
psycopg2==2.7.3.1
pylint==1.7.2
pyshorteners==0.6.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
singledispatch==3.4.0.3
six==1.10.0
social-auth-core==1.7.0
social-auth-app-django==2.1.0
tweepy==3.5.0
urllib3==1.22
vine==1.1.4
whitenoise==3.3.0
wrapt==1.10.11
